requirements.txt
azure-functions
dnspython==1.16.0
pymongo
